Remove ID to position mapping in the database
As long as we use a numeric value for the ID that can be transformed
directly into `u32` (used by the bitmap inside the index), we
don't need an extra DB entry. This reduces heavily the load when
reading the data from the DB, as we don't need to do two operations
(read ID by position + read data), but just the latest.