Define a query delta scope
A delta scope defines a collection of deltas under a certain scope,
so that deltas can be organized depending on their context.

`context` is defined as a `u64` identifier.
`deltas` are stored by context now, and each context has a `BTreeMap`
of deltas by timestamp so that range operations are still possible.